Use double data type for the SPM atlas affine
UniSeg errored for me because the input affine I gave was torch32. think it is best to use double data type for the SPM atlas affine matrix because it is what is returned by default when reading with io.map().affine, plus it is better for numerical precision without much memory overhead.