add initial Dockerfile
This dockerfile builds nitorch and compiles the extensions. The first
stage builds the (compiled) wheel. The second build stage installs that
wheel. The dockerfile bootstraps pytorch's official docker image. The
first stage uses their "devel" image and the second image uses
"runtime".

We should test that nitorch is usable and actually uses the compiled
things. It is possible we will need to update the `LD_LIBRARY_PATH` in
the second build stage.